69. Setup Custom Surface Types  &  70. Add Multiple Hit Effect Types
This is useful if you want to play different sounds depending on the surface hit, or if you want to play an effect if you hit a headshot.

1. Add Surface Types in the Project Properties

2. Restart UE4 Editor for new surface types to set in.

3. Create a Physics Matterial in the content editor

4. Select the correct surface type you want to useful

5. Apply the Physics Matterial to the object or Physics Asset part on a skeletal mesh. (like the head)

6. Setup Mutiple Hit Effects by using:

EPhysicalSurface HitSurfaceType = UPhysicalMaterial::DetermineSurfaceType(Hit.PhysMaterial.Get());                   // Used Get() here, because its a weak object pointer and we want to get the object.

7. If you used a linetrace or collision hit check to get the other object, make sure you add QueryParams.bReturnPhysicalMaterial = true; for the linetrace / hit, or else the value will be blank when you try step 6.